% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior}
\alias{posterior}
\title{Posterior Estimates}
\usage{
posterior(obj, burnin = 100)
}
\arguments{
\item{obj}{object of class \code{"zinLDA_gibbs"}.}

\item{burnin}{number of burn-in runs of the MCMC.}
}
\value{
A list with the following elements:
\item{delta}{posterior indicator of structural zeros.}
\item{beta}{posterior probabilities of each taxon belonging to each subcommunity.}
\item{theta}{posterior probabilities of each subcommunity in each sample.}
}
\description{
Calculate the posterior estimates of \code{delta}, \code{beta}, and \code{theta} for a fitted zinLDA model.
}
