% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdu.R
\name{rdu}
\alias{rdu}
\title{The Discrete Uniform Distribution}
\usage{
rdu(n, min, max)
}
\arguments{
\item{n}{number of observations.}

\item{min}{lower limit of the distribution.}

\item{max}{upper limit of the distribution.}
}
\value{
A length n numeric vector of realizations from a DU(\code{min}, \code{max}) distribution.
}
\description{
\code{rdu} generates a random sample of size \code{n} from a discrete uniform
distribution on the interval \code{min} to \code{max}.
}
\examples{
N = rdu(n = 50, min = 10000, max = 15000)
}
