% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateZINLDA.R
\name{simulateZINLDA}
\alias{simulateZINLDA}
\title{zinLDA Simulated Data}
\usage{
simulateZINLDA(D, V, N, K, Alpha, Pi, a, b)
}
\arguments{
\item{D}{number of samples.}

\item{V}{number of unique taxa.}

\item{N}{vector of length D containing the total number of sequencing readings per sample.}

\item{K}{number of latent subcommunities.}

\item{Alpha}{scalar symmetric hyperparameter of the Dirichlet prior on theta.}

\item{Pi}{scalar symmetric zero-inflation hyperparameter of the ZIGD prior on beta.}

\item{a}{scalar symmetric hyperparameter of the ZIGD prior on beta.}

\item{b}{scalar symmetric hyperparameter of the ZIGD prior on beta.}
}
\value{
A list containing the following elements:
\item{cohort}{D-length list of character vectors containing the taxa assigned
to each sequencing read in each microbial sample.}
\item{z}{D-length list of vectors containing the subcommunity assignments for
each sequencing read in each microbial sample.}
\item{sampleTaxaMatrix}{matrix of counts, analogous to an OTU matrix.}
\item{theta}{matrix of subcommunity probabilities per sample.}
\item{beta}{matrix of taxa probabilities per subcommunity.}
\item{delta}{matrix of structural zero indicators for each taxa and subcommunity.}
}
\description{
\code{simulateZINLDA} generates sparse count data according to a zero-inflated latent Dirichlet allocation model.
}
\examples{
N.d = rdu(n = 50, min = 10000, max = 15000)
simData = simulateZINLDA(D = 50, V = 100, N = N.d, K = 5, Alpha = .1,
                         Pi = 0.4, a = .05, b = 10)
}
