% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinLDA.R
\name{zinLDA}
\alias{zinLDA}
\title{Zero-Inflated Latent Dirichlet Allocation}
\usage{
zinLDA(sampleTaxaMatrix, K, alpha, pi, a, b, runs = 500)
}
\arguments{
\item{sampleTaxaMatrix}{object of class \code{matrix} with integer entries.}

\item{K}{number of latent subcommunities.}

\item{alpha}{positive scalar symmetric hyperparameter of the Dirichlet prior on theta.}

\item{pi}{scalar symmetric zero-inflation probability of the ZIGD prior on beta.}

\item{a}{positive scalar symmetric hyperparameter of the ZIGD prior on beta.}

\item{b}{positive scalar symmetric hyperparameter of the ZIGD prior on beta.}

\item{runs}{length of the MCMC chain.}
}
\value{
An object of class \code{"zinLDA_gibbs"} containing parameter estimates
of \code{theta}, \code{beta}, and \code{delta} from each of the MCMC runs.
}
\description{
Estimate a zinLDA model using Gibbs Sampling.
}
